% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{expandMat}
\alias{expandMat}
\title{Expand a matrix of predictors to a matrix of latent variables.}
\usage{
expandMat(X, group)
}
\arguments{
\item{X}{Design matrix (features).}

\item{group}{A list of vectors containing group information.}
}
\value{
A sparse matrix with all latent variables included.
}
\description{
Expand a matrix of predictors to a matrix of latent variables.
}
