% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{incidenceMat}
\alias{incidenceMat}
\title{Create an incidence matrix indicating group membership}
\usage{
incidenceMat(X, group)
}
\arguments{
\item{X}{Design matrix (features).}

\item{group}{A list of vectors containing group information.}
}
\value{
A sparse matrix with all latent variables included.
}
\description{
Create an incidence matrix indicating group membership
}
