% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cv.ovganet}
\alias{plot.cv.ovganet}
\title{Plot method for cv.ovganet}
\usage{
\method{plot}{cv.ovganet}(obj, ...)
}
\arguments{
\item{obj}{An object of class 'cv.ovganet'}

\item{...}{Additional arguments to be passed to the plot function of 'plot.cv.oem'}
}
\description{
Plot method for cv.ovganet
}
