% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ovganet}
\alias{plot.ovganet}
\title{Plot method for ovganet plotting (coef. path)}
\usage{
\method{plot}{ovganet}(obj, ...)
}
\arguments{
\item{obj}{An object of class 'ovganet'}

\item{...}{Additional arguments passed to the plot function of 'plot.oem'}
}
\description{
Plot method for ovganet plotting (coef. path)
}
