% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cv.ovganet}
\alias{predict.cv.ovganet}
\title{Predict method for cv.ovganet}
\usage{
\method{predict}{cv.ovganet}(object, newx, ...)
}
\arguments{
\item{object}{An object of class 'cv.ovganet'}

\item{newx}{New data to predict from. If not provided, will use the data from the original fit.}

\item{...}{Additional arguments passed to the predict function of 'predict.cv.oem'}
}
\value{
Predicted values based on the 'cv.ovganet' object
}
\description{
Predict method for cv.ovganet
}
