% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ovganet}
\alias{predict.ovganet}
\title{Predict method for ovganet}
\usage{
\method{predict}{ovganet}(obj, newx, ...)
}
\arguments{
\item{obj}{An object of class 'ovganet'}

\item{newx}{New data to predict from. If not provided, will use the data from the original fit.}

\item{...}{Additional arguments to be passed to the predict function of 'predict.oem'}
}
\value{
Predicted values based on the 'ovganet' object
}
\description{
Predict method for ovganet
}
