% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cv.ovganet}
\alias{summary.cv.ovganet}
\title{Summary method for cv.ovganet}
\usage{
summary.cv.ovganet(object, ...)
}
\arguments{
\item{object}{An object of class 'cv.ovganet'}

\item{...}{Additional arguments passed to the summary function of 'summary.cv.oem'}
}
\value{
A summary object for 'cv.ovganet'
}
\description{
Summary method for cv.ovganet
}
