% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ovganet}
\alias{summary.ovganet}
\title{Summary method for ovganet}
\usage{
summary.ovganet(object, ...)
}
\arguments{
\item{object}{An object of class 'ovganet'}

\item{...}{Additional arguments passed to the summary function of 'summary.oem'}
}
\value{
A summary object for 'ovganet'
}
\description{
Summary method for ovganet
}
