% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.ovganet.R
\name{cv.ovganet}
\alias{cv.ovganet}
\title{Cross-validated overlapping group elastic net using package 'oem'}
\usage{
cv.ovganet(
  X,
  y,
  group,
  weights = NULL,
  alpha = 1,
  type.measure = "deviance",
  family = c("gaussian", "binomial"),
  ...
)
}
\arguments{
\item{X}{Design matrix (features). Note that "oem" is optimized for n » p settings.}

\item{y}{Response vector (outcomes).}

\item{group}{A list of vectors containing group information.}

\item{weights}{A vector of weights for each group.}

\item{alpha}{Mixing value for elastic.net.}

\item{type.measure}{Measure to evaluate for cross-validation.
The default is type.measure = "deviance." See package "oem" for more options.}

\item{family}{Use "gaussian" for least squares problems and "binomial" for binary response.}

\item{...}{other parameters to be passed to "cv.ovganet" function.}
}
\value{
An object with S3 class "cv.oem" (same as cv.oem output).
}
\description{
Cross-validated overlapping group elastic net using package 'oem'
}
\examples{
To be filled in.
}
