% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.ovganet.R
\name{cv.ovganet}
\alias{cv.ovganet}
\title{Cross-validated overlapping group elastic net using package 'oem'}
\usage{
cv.ovganet(
  X,
  y,
  group,
  weights = NULL,
  alpha = 1,
  type.measure = "deviance",
  family = c("gaussian", "binomial"),
  ...
)
}
\arguments{
\item{X}{Design matrix (features). Note that the "oem" package we use is optimized for n >> p settings.}

\item{y}{Response vector (outcomes).}

\item{group}{A list of vectors containing group information.}

\item{weights}{A vector of weights for each group.}

\item{alpha}{Mixing value for elastic.net.}

\item{type.measure}{Measure to evaluate for cross-validation.
The default is type.measure = "deviance." See package "oem" for more options.}

\item{family}{Use "gaussian" for least squares problems and "binomial" for binary response.}

\item{...}{other parameters passed to "cv.ovganet" function.}
}
\value{
An object with S3 class "cv.ovganet".
}
\description{
Cross-validated overlapping group elastic net using package 'oem'
}
\examples{
library(ovganet)
data(mtcars)
X <- as.matrix(mtcars[,-1])
y <- as.vector(mtcars$mpg)
group = list(c(1,2), c(2,3), c(3,4,5),c(4,5,6))
cvfit <- cv.ovganet(X = X, y = y, group = group, family ='gaussian')
}
