% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovganet.R
\name{ovganet}
\alias{ovganet}
\title{Overlapping group elastic net using package 'oem'}
\usage{
ovganet(
  X,
  y,
  group,
  weights = NULL,
  alpha = 1,
  family = c("gaussian", "binomial"),
  ...
)
}
\arguments{
\item{X}{Design matrix (features). Note that "oem" is optimized for n » p settings.}

\item{y}{Response vector (outcomes).}

\item{group}{A list of vectors containing group information.}

\item{weights}{A vector of weights for each group.}

\item{alpha}{Mixing value for elastic.net.}

\item{family}{Use "gaussian" for least squares problems and "binomial" for binary response.}

\item{...}{other parameters to be passed to "cv.ovganet" function.}
}
\value{
An object with S3 class "ovganet".
}
\description{
Overlapping group elastic net using package 'oem'
}
\examples{
library(ovganet)
overlap_grp_lasso <- ovganet(X = X, y = y, group = group, weights = group_weights, family ='binomial')
}
