% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cv.ovganet}
\alias{summary.cv.ovganet}
\title{Summary method for cv.ovganet}
\usage{
summary.cv.ovganet(obj, ...)
}
\arguments{
\item{obj}{An obj of class 'cv.ovganet'}

\item{...}{Additional arguments to be passed to the summary function of 'summary.cv.oem'}
}
\value{
A summary obj for 'cv.ovganet'
}
\description{
Summary method for cv.ovganet
}
